requirements.txt
pydantic~=2.11.7
fastapi~=0.116.0
python-dotenv~=1.1.1